+++ .gitignore
@@ -4,3 +4,4 @@ data/*
 chroma
 __pycache__/statuteparser.cpython-310.pyc
 docs/statutes/cache/*
+*.pyc
